Revert "Revert "fix(ai): auto attachments (#624)" (#683)"
This reverts commit 564af4e0cd8c2d016581b1e5d8641e6cbcd8d1fa.
